chore: move L1 Messages lib into client space
(cherry picked from commit 58caefa593515777c2e787c7d0466df005f36019)
