+++ crates/client/l1-messages/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "mc-l1-messages"
+version = "0.1.0"
+description = "L1 Messages processing library."
+homepage = "https://github.com/keep-starknet-strange/madara"
+edition = "2021"
+license = "MIT"
+publish = false
+repository = "https://github.com/keep-starknet-strange/madara"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+log = "0.4.19"
+
+# Madara Runtime
+madara-runtime = { workspace = true }
+
+# Starknet dependencies
+starknet_api = { workspace = true }
+
+# Primitives
+mp-felt = { workspace = true }
+
+# Madara Primitives
+mp-transactions = { workspace = true }
+
+# Substrate Primitives
+sp-api = { workspace = true }
+sp-core = { workspace = true }
+sp-runtime = { workspace = true }
+
+# Substrate client dependencies
+sc-client-api = { workspace = true }
+sc-transaction-pool-api = { workspace = true }
+
+# Other third party dependencies
+ethers = { workspace = true }
+parity-scale-codec = { workspace = true }
+
