+++ crates/client/l1-messages/src/lib.rs
@@ -1,3 +1,4 @@
-pub mod contract;
 pub mod error;
 pub mod worker;
+
+mod contract;
