+++ crates/node/src/service.rs
@@ -19,6 +19,7 @@ use mc_data_availability::celestia::CelestiaClient;
 use mc_data_availability::ethereum::config::EthereumConfig;
 use mc_data_availability::ethereum::EthereumClient;
 use mc_data_availability::{DaClient, DaLayer, DataAvailabilityWorker};
+use mc_l1_messages::worker::L1MessagesWorkerConfig;
 use mc_mapping_sync::MappingSyncWorker;
 use mc_storage::overrides_handle;
 use mp_sequencer_address::{
@@ -42,8 +43,6 @@ use sp_consensus_aura::sr25519::AuthorityPair as AuraPair;
 use sp_offchain::STORAGE_PREFIX;
 
 use crate::genesis_block::MadaraGenesisBlockBuilder;
-use crate::l1_messages;
-use crate::l1_messages::worker::L1MessagesWorkerConfig;
 use crate::rpc::StarknetDeps;
 use crate::starknet::{db_config_dir, MadaraBackend};
 
@@ -585,7 +584,7 @@ pub fn new_full(
     task_manager.spawn_essential_handle().spawn(
         "L1 Messages",
         Some(MADARA_TASK_GROUP),
-        l1_messages::worker::run_worker(L1MessagesWorkerConfig::default(), client, transaction_pool, db_backend),
+        mc_l1_messages::worker::run_worker(L1MessagesWorkerConfig::default(), client, transaction_pool, db_backend),
     );
 
     network_starter.start_network();
